# backend/API/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: |-
          The HTTP status code of the response.
          example: 400
        type: integer
      error:
        description: |-
          The standard HTTP status text for the code.
          example: Bad Request
        type: string
      message:
        description: |-
          A user-friendly message describing the specific issue.
          example: Invalid request body or validation failed.
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the social-network application.
  title: Social-network Application API
  version: "1.0"
paths:
...
  /api/v1/my-commented-posts:
    get:
      description: Retrieves a list of posts that the current user has interacted
        with via comments. Useful for activity history.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get posts commented on
      tags:
      - User Activity
  /api/v1/my-posts:
    get:
      description: Retrieves a full list of posts created by the logged-in user, including
        nested comments and reactions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Failed to load posts or associated data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get own posts
      tags:
      - User Activity
...
swagger: "2.0"